+++ sdk/core/System.ClientModel/src/Options/ClientLoggingOptions.cs
@@ -4,6 +4,7 @@
 using System.ClientModel.Internal;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Threading;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -64,31 +65,6 @@ public class ClientLoggingOptions
     }
 
     /// <summary>
-    /// Initializes a new mutable instance of <see cref="ClientLoggingOptions"/>
-    /// from an existing instance.  This constructor can be used to create a
-    /// mutable copy of an instance that may have been frozen.
-    /// </summary>
-    /// <param name="options">The <see cref="ClientLoggingOptions"/> to copy.</param>
-    public ClientLoggingOptions(ClientLoggingOptions options)
-    {
-        Argument.AssertNotNull(options, nameof(options));
-
-        _enableLogging = options._enableLogging;
-        _enableMessageLogging = options._enableMessageLogging;
-        _enableMessageContentLogging = options._enableMessageContentLogging;
-        _messageContentSizeLimit = options._messageContentSizeLimit;
-        _loggerFactory = options._loggerFactory;
-        if (options._allowedHeaderNames is not null)
-        {
-            _allowedHeaderNames = new ChangeTrackingStringList(options._allowedHeaderNames);
-        }
-        if (options._allowedQueryParameters is not null)
-        {
-            _allowedQueryParameters = new ChangeTrackingStringList(options._allowedQueryParameters);
-        }
-    }
-
-    /// <summary>
     /// Initializes a new instance of <see cref="ClientLoggingOptions"/> from configuration.
     /// </summary>
     /// <param name="section">The configuration section to bind from.</param>
@@ -130,10 +106,9 @@ public class ClientLoggingOptions
     /// Options become read-only when the <see cref="ClientPipeline"/> they are
     /// associated with is created, or when <see cref="Freeze"/> is called
     /// explicitly.  To create a mutable copy of a read-only instance, use
-    /// the <see cref="ClientLoggingOptions(ClientLoggingOptions)"/> copy
-    /// constructor.
+    /// <see cref="Clone"/>.
     /// </remarks>
-    public bool IsReadOnly => Volatile.Read(_frozen);
+    public bool IsReadOnly => Volatile.Read(ref _frozen);
 
     /// <summary>
     /// Gets or sets the implementation of <see cref="ILoggerFactory"/> to use to
@@ -297,6 +272,40 @@ public class ClientLoggingOptions
         {
             _allowedQueryParameters.Freeze();
         }
+    }
+
+    /// <summary>
+    /// Creates a new mutable instance of <see cref="ClientLoggingOptions"/> from this
+    /// instance.  This method can be used to create a mutable copy of an instance that
+    /// may have been frozen.
+    /// </summary>
+    /// <returns>A new mutable <see cref="ClientLoggingOptions"/> with the same settings
+    /// as this instance.</returns>
+    public virtual ClientLoggingOptions Clone()
+    {
+        var clone = new ClientLoggingOptions();
+        clone._enableLogging = _enableLogging;
+        clone._enableMessageLogging = _enableMessageLogging;
+        clone._enableMessageContentLogging = _enableMessageContentLogging;
+        clone._messageContentSizeLimit = _messageContentSizeLimit;
+        clone._loggerFactory = _loggerFactory;
+        if (_allowedHeaderNames is not null && _allowedHeaderNames.HasChanged)
+        {
+            clone._allowedHeaderNames = new ChangeTrackingStringList();
+            foreach (string item in _allowedHeaderNames)
+            {
+                clone._allowedHeaderNames.Add(item);
+            }
+        }
+        if (_allowedQueryParameters is not null && _allowedQueryParameters.HasChanged)
+        {
+            clone._allowedQueryParameters = new ChangeTrackingStringList();
+            foreach (string item in _allowedQueryParameters)
+            {
+                clone._allowedQueryParameters.Add(item);
+            }
+        }
+        return clone;
     }
 
     /// <summary>